+++ app/Models/Audiovisual.php
@@ -22,7 +22,7 @@ If not, see <https://www.gnu.org/licenses/>.
 
 namespace App\Models;
 
-use App\Traits\CommentTrait;
+use App\Contracts\HasComments;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
@@ -38,9 +38,9 @@ use Spatie\Translatable\HasTranslations;
  * @property bool $is_translated
  * @property int|null $media_id Runtime-Zuweisung im ProjectController, nicht DB-Spalte.
  */
-class Audiovisual extends Model
+class Audiovisual extends Model implements HasComments
 {
-    use CommentTrait, HasFactory, HasTranslations, LogsActivity, SoftDeletes;
+    use HasFactory, HasTranslations, LogsActivity, SoftDeletes;
 
     /**
      * The attributes that are mass assignable.
@@ -53,10 +53,8 @@ class Audiovisual extends Model
 
     /**
      * Get all comments
-     *
-     * @return MorphMany
      */
-    public function comments()
+    public function comments(): MorphMany
     {
         return $this->morphMany(Comment::class, 'commentable')->whereNull('parent_id');
     }