+++ app/Models/Entry.php
@@ -22,7 +22,7 @@ If not, see <https://www.gnu.org/licenses/>.
 
 namespace App\Models;
 
-use App\Traits\CommentTrait;
+use App\Contracts\HasComments;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -44,9 +44,9 @@ use Spatie\Translatable\HasTranslations;
  * @property Collection<int, MediaContent> $mediaContent
  * @property mixed $media Runtime-Zuweisung in ProjectController::allData (Media-Snapshot je Entry), nicht DB-Spalte.
  */
-class Entry extends Model
+class Entry extends Model implements HasComments
 {
-    use CommentTrait, HasFactory, HasTranslations, LogsActivity, SoftDeletes;
+    use HasFactory, HasTranslations, LogsActivity, SoftDeletes;
 
     /**
      * The attributes that are mass assignable.
@@ -127,10 +127,8 @@ class Entry extends Model
 
     /**
      * Get all comments
-     *
-     * @return MorphMany
      */
-    public function comments()
+    public function comments(): MorphMany
     {
         return $this->morphMany(Comment::class, 'commentable')->whereNull('parent_id');
     }